extract CoastlineMenu component and add test
