+++ forest_lite/client/src/CoastlineMenu.css
@@ -1,0 +1,13 @@
+fieldset {
+    font-family: Helvetica, Arial, sans-serif;
+    margin: 4px;
+    padding: 4px;
+    border: none;
+}
+
+.label {
+    font-weight: bold;
+    font-family: Helvetica, Arial, sans-serif;
+    margin-top: 20px;
+    margin-left: 6px;
+}
