+++ forest_lite/client/src/Sidebar.css
@@ -1,18 +1,3 @@
-fieldset {
-    font-family: Helvetica, Arial, sans-serif;
-    margin: 4px;
-    padding: 4px;
-    border: none;
-}
-
-.label {
-    font-weight: bold;
-    font-family: Helvetica, Arial, sans-serif;
-    margin-top: 20px;
-    margin-left: 6px;
-}
-
-
 .layer-menu-container {
     background-color: white;
     margin: 10px;
